ball.setVelocity, correct redraw of ball and remove uneccesary calculation
ball.setVelocity() will change the speed of the ball in real time.

When we drew the ball we drew with uneven pixels. I round the positions and dimensions to whole numbers and draw the ball

When we hit a wall, we do not need to recalculate K. Instead, we just flip either dX or dY (depending on the which wall we hit).